+++ lib/mbti_selection_page.dart
@@ -1,6 +1,8 @@
 import 'package:flutter/material.dart';
 import 'calendar_selection_page.dart'; // 달력 선택 페이지 파일을 import
 import 'loginpage.dart'; // 로그인 페이지 파일을 import
+import 'BestCourseTop3.dart'; // 새로운 파일 import
+import 'GyeongNamRecommend.dart'; // 새로운 파일 import
 
 class MBTISelectionApp extends StatelessWidget {
   @override
@@ -76,15 +78,19 @@ class _MBTISelectionPageState extends State<MBTISelectionPage> {
 
   @override
   Widget build(BuildContext context) {
+    final double screenWidth = MediaQuery.of(context).size.width;
+    final double screenHeight = MediaQuery.of(context).size.height;
+
     return Scaffold(
       backgroundColor: Colors.white,
       appBar: PreferredSize(
-        preferredSize: Size.fromHeight(60.0),
+        preferredSize: Size.fromHeight(screenHeight * 0.08),
         child: Container(
           color: Colors.white,
           child: AppBar(
             centerTitle: true,
-            title: Image.asset('assets/mbtilogo.jpg', height: 40), // MBTI 로고 경로
+            title: Image.asset('assets/mbtilogo.jpg',
+                height: screenHeight * 0.05), // MBTI 로고 경로
             backgroundColor: Colors.white,
             elevation: 0,
             actions: [
@@ -105,15 +111,18 @@ class _MBTISelectionPageState extends State<MBTISelectionPage> {
               thickness: 1,
               height: 1,
             ),
-            SizedBox(height: 20),
+            SizedBox(height: screenHeight * 0.03),
             Row(
               children: [
-                SizedBox(width: 20),
-                Image.asset('assets/animation.png', height: 50), // 사람 아이콘 경로
-                SizedBox(width: 10),
+                SizedBox(width: screenWidth * 0.05),
+                Image.asset('assets/animation.png',
+                    height: screenHeight * 0.07), // 사람 아이콘 경로
+                SizedBox(width: screenWidth * 0.02),
                 Expanded(
                   child: Container(
-                    padding: EdgeInsets.symmetric(vertical: 10, horizontal: 20),
+                    padding: EdgeInsets.symmetric(
+                        vertical: screenHeight * 0.015,
+                        horizontal: screenWidth * 0.05),
                     decoration: BoxDecoration(
                       color: Colors.grey[200],
                       borderRadius: BorderRadius.circular(20),
@@ -122,17 +131,18 @@ class _MBTISelectionPageState extends State<MBTISelectionPage> {
                       selectedMBTI.length == 4
                           ? '${selectedMBTI.join()} OOO님!\n오늘은 경상남도 어디로 떠나볼까요?'
                           : '경상남도 추천지를 원하시면 MBTI를 선택해주세요 !!',
-                      style: TextStyle(fontSize: 16),
+                      style: TextStyle(fontSize: screenHeight * 0.02),
                     ),
                   ),
                 ),
-                SizedBox(width: 20),
+                SizedBox(width: screenWidth * 0.05),
               ],
             ),
-            SizedBox(height: 20),
+            SizedBox(height: screenHeight * 0.02),
             Text(
               'MBTI 맞춤형 간단 추천 코스',
-              style: TextStyle(fontSize: 20, fontWeight: FontWeight.bold),
+              style: TextStyle(
+                  fontSize: screenHeight * 0.03, fontWeight: FontWeight.bold),
               textAlign: TextAlign.center,
             ),
             StepOneWidget(
@@ -140,24 +150,10 @@ class _MBTISelectionPageState extends State<MBTISelectionPage> {
               toggleSelection: toggleSelection,
               onNextPressed: navigateToCalendarPage,
             ),
-            SizedBox(height: 20), // 추가 여백
-            RecommendedCourses(),
-            SizedBox(height: 20), // 추가 여백
-            Padding(
-              padding: const EdgeInsets.symmetric(horizontal: 20.0),
-              child: Row(
-                mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                children: [
-                  Text(
-                    '경상남도 행사 & 축제',
-                    style: TextStyle(fontSize: 20, fontWeight: FontWeight.bold),
-                  ),
-                  SortOptions(),
-                ],
-              ),
-            ),
-            EventsAndFestivals(),
-            SizedBox(height: 20), // 추가 여백
+            SizedBox(height: screenHeight * 0.02), // 추가 여백
+            BestCourseTop3(), // 인기 추천코스 BEST 3 위젯 사용
+            SizedBox(height: screenHeight * 0.02), // 추가 여백
+            GyeongNamRecommend(), // 경상남도 행사 & 축제 위젯 사용
           ],
         ),
       ),
@@ -165,17 +161,18 @@ class _MBTISelectionPageState extends State<MBTISelectionPage> {
         backgroundColor: Colors.white, // 하단바 배경색을 흰색으로 설정
         items: [
           BottomNavigationBarItem(
-            icon: Image.asset('assets/home.jpg', height: 30), // 홈 아이콘 경로
+            icon: Image.asset('assets/home.jpg',
+                height: screenHeight * 0.04), // 홈 아이콘 경로
             label: '홈',
           ),
           BottomNavigationBarItem(
-            icon:
-                Image.asset('assets/location1.png', height: 30), // 여행 일정 아이콘 경로
+            icon: Image.asset('assets/location1.png',
+                height: screenHeight * 0.04), // 여행 일정 아이콘 경로
             label: '여행 일정',
           ),
           BottomNavigationBarItem(
-            icon:
-                Image.asset('assets/myprofile.png', height: 30), // 내 정보 아이콘 경로
+            icon: Image.asset('assets/myprofile.png',
+                height: screenHeight * 0.04), // 내 정보 아이콘 경로
             label: '내 정보',
           ),
         ],
@@ -199,10 +196,12 @@ class StepOneWidget extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    final double screenHeight = MediaQuery.of(context).size.height;
+
     return Padding(
-      padding: const EdgeInsets.all(20.0),
+      padding: EdgeInsets.all(screenHeight * 0.03),
       child: Container(
-        padding: EdgeInsets.all(20),
+        padding: EdgeInsets.all(screenHeight * 0.03),
         decoration: BoxDecoration(
           color: Colors.grey[200],
           borderRadius: BorderRadius.circular(10),
@@ -212,14 +211,14 @@ class StepOneWidget extends StatelessWidget {
           children: [
             Text(
               'STEP 01 | MBTI 입력',
-              style: TextStyle(fontSize: 16),
+              style: TextStyle(fontSize: screenHeight * 0.02),
               textAlign: TextAlign.left,
             ),
-            SizedBox(height: 10),
+            SizedBox(height: screenHeight * 0.01),
             Wrap(
               alignment: WrapAlignment.center,
-              spacing: 10,
-              runSpacing: 10,
+              spacing: screenHeight * 0.01,
+              runSpacing: screenHeight * 0.01,
               children: [
                 MBTIGroup(
                   verticalItems: ['E', 'I'],
@@ -243,14 +242,14 @@ class StepOneWidget extends StatelessWidget {
                 ),
               ],
             ),
-            SizedBox(height: 20),
+            SizedBox(height: screenHeight * 0.03),
             ElevatedButton(
               onPressed: selectedMBTI.length == 4 ? onNextPressed : null,
               child: Text('다음'),
               style: ElevatedButton.styleFrom(
                 backgroundColor: Colors.grey[800],
                 foregroundColor: Colors.white,
-                padding: EdgeInsets.symmetric(vertical: 15),
+                padding: EdgeInsets.symmetric(vertical: screenHeight * 0.02),
               ),
             ),
           ],
@@ -273,12 +272,14 @@ class MBTIGroup extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    final double screenHeight = MediaQuery.of(context).size.height;
+
     return Container(
       decoration: BoxDecoration(
         color: Colors.white,
         borderRadius: BorderRadius.circular(10),
       ),
-      padding: EdgeInsets.all(5),
+      padding: EdgeInsets.all(screenHeight * 0.01),
       child: Column(
         children: verticalItems.map((mbti) {
           return MBTIButton(
@@ -305,13 +306,15 @@ class MBTIButton extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
+    final double screenHeight = MediaQuery.of(context).size.height;
+
     return Container(
-      width: 50,
-      height: 50,
-      margin: EdgeInsets.all(3),
+      width: screenHeight * 0.07,
+      height: screenHeight * 0.07,
+      margin: EdgeInsets.all(screenHeight * 0.004),
       child: ElevatedButton(
         onPressed: () => toggleSelection(mbti),
-        child: Text(mbti, style: TextStyle(fontSize: 18)),
+        child: Text(mbti, style: TextStyle(fontSize: screenHeight * 0.025)),
         style: ElevatedButton.styleFrom(
           backgroundColor: isSelected ? Colors.grey[300] : Colors.white,
           foregroundColor: Colors.black,
@@ -320,268 +323,6 @@ class MBTIButton extends StatelessWidget {
             side: BorderSide(color: Colors.grey),
           ),
           elevation: 1,
-        ),
-      ),
-    );
-  }
-}
-
-class RecommendedCourses extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.stretch,
-      children: [
-        Padding(
-          padding: const EdgeInsets.symmetric(horizontal: 20.0),
-          child: Text(
-            '인기 추천코스\nBEST 3',
-            style: TextStyle(fontSize: 20, fontWeight: FontWeight.bold),
-          ),
-        ),
-        SizedBox(height: 10),
-        Container(
-          height: 200,
-          child: ListView(
-            scrollDirection: Axis.horizontal,
-            children: [
-              RecommendedCourseCard(
-                imagePath: 'assets/course1.png',
-                title: '해금강해안도로',
-                location: '거제',
-                mbti: 'ENTP',
-              ),
-              RecommendedCourseCard(
-                imagePath: 'assets/course2.png',
-                title: '시간여행마을',
-                location: '창녕',
-                mbti: 'ISTJ',
-              ),
-              RecommendedCourseCard(
-                imagePath: 'assets/course3.png',
-                title: '우포늪',
-                location: '창녕',
-                mbti: 'ESTJ',
-              ),
-              Container(
-                width: 100,
-                alignment: Alignment.center,
-                child: TextButton(
-                  onPressed: () {},
-                  child: Row(
-                    mainAxisAlignment: MainAxisAlignment.center,
-                    children: [
-                      Image.asset(
-                        'assets/your_image.png', // 여기에 이미지 경로를 넣으세요
-                        height: 30, // 적절한 높이로 조절하세요
-                      ),
-                      SizedBox(width: 10), // 이미지와 텍스트 사이의 간격을 조절하세요
-                      Text(
-                        '다른 코스가 보고 싶다면?',
-                        style: TextStyle(color: Colors.black),
-                        textAlign: TextAlign.center,
-                      ),
-                    ],
-                  ),
-                ),
-              ),
-            ],
-          ),
-        ),
-      ],
-    );
-  }
-}
-
-class RecommendedCourseCard extends StatelessWidget {
-  final String imagePath;
-  final String title;
-  final String location;
-  final String mbti;
-
-  const RecommendedCourseCard({
-    required this.imagePath,
-    required this.title,
-    required this.location,
-    required this.mbti,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      width: 160,
-      margin: EdgeInsets.only(left: 20, right: 10),
-      decoration: BoxDecoration(
-        borderRadius: BorderRadius.circular(10),
-        image: DecorationImage(
-          image: AssetImage(imagePath),
-          fit: BoxFit.cover,
-        ),
-      ),
-      child: Stack(
-        children: [
-          Positioned(
-            bottom: 10,
-            left: 10,
-            child: Text(
-              '$mbti\n$title\n$location',
-              style: TextStyle(
-                color: Colors.white,
-                fontSize: 16,
-                fontWeight: FontWeight.bold,
-                shadows: [
-                  Shadow(
-                    offset: Offset(0, 1),
-                    blurRadius: 2,
-                    color: Colors.black.withOpacity(0.5),
-                  ),
-                ],
-              ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class EventsAndFestivals extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      crossAxisAlignment: CrossAxisAlignment.stretch,
-      children: [
-        SizedBox(height: 10),
-        Container(
-          height: 200,
-          child: ListView(
-            scrollDirection: Axis.horizontal,
-            children: [
-              EventFestivalCard(
-                imagePath: 'assets/event1.png',
-                title: '거제 바다로 세계로',
-                description: '거제의 대표적인 여름 축제',
-              ),
-              EventFestivalCard(
-                imagePath: 'assets/event2.png',
-                title: '창녕 남지 유채꽃 축제',
-                description: '봄에 즐기는 아름다운 유채꽃',
-              ),
-            ],
-          ),
-        ),
-      ],
-    );
-  }
-}
-
-class EventFestivalCard extends StatelessWidget {
-  final String imagePath;
-  final String title;
-  final String description;
-
-  const EventFestivalCard({
-    required this.imagePath,
-    required this.title,
-    required this.description,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Container(
-      width: 160,
-      margin: EdgeInsets.only(left: 20, right: 10),
-      decoration: BoxDecoration(
-        borderRadius: BorderRadius.circular(10),
-        image: DecorationImage(
-          image: AssetImage(imagePath),
-          fit: BoxFit.cover,
-        ),
-      ),
-      child: Stack(
-        children: [
-          Positioned(
-            bottom: 10,
-            left: 10,
-            child: Text(
-              '$title\n$description',
-              style: TextStyle(
-                color: Colors.white,
-                fontSize: 16,
-                fontWeight: FontWeight.bold,
-                shadows: [
-                  Shadow(
-                    offset: Offset(0, 1),
-                    blurRadius: 2,
-                    color: Colors.black.withOpacity(0.5),
-                  ),
-                ],
-              ),
-            ),
-          ),
-        ],
-      ),
-    );
-  }
-}
-
-class SortOptions extends StatefulWidget {
-  @override
-  _SortOptionsState createState() => _SortOptionsState();
-}
-
-class _SortOptionsState extends State<SortOptions> {
-  String selectedOption = 'latest';
-
-  @override
-  Widget build(BuildContext context) {
-    return Row(
-      children: [
-        SortOption(
-          text: '최신순',
-          isSelected: selectedOption == 'latest',
-          onTap: () {
-            setState(() {
-              selectedOption = 'latest';
-            });
-          },
-        ),
-        SizedBox(width: 10),
-        SortOption(
-          text: '인기순',
-          isSelected: selectedOption == 'popular',
-          onTap: () {
-            setState(() {
-              selectedOption = 'popular';
-            });
-          },
-        ),
-      ],
-    );
-  }
-}
-
-class SortOption extends StatelessWidget {
-  final String text;
-  final bool isSelected;
-  final VoidCallback onTap;
-
-  const SortOption({
-    required this.text,
-    required this.isSelected,
-    required this.onTap,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return GestureDetector(
-      onTap: onTap,
-      child: Text(
-        text,
-        style: TextStyle(
-          fontSize: 16,
-          fontWeight: FontWeight.bold,
-          color: isSelected ? Colors.black : Colors.grey,
         ),
       ),
     );
